+++ src/classes/Cell.ts
