+++ src/classes/Circuit.ts
