+++ src/classes/TorSocket.ts
